requirements.txt
aiofiles==23.2.1
aiogram==3.5.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
amplitude-analytics==1.1.4
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
backports.tarfile==1.1.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.6
decorator==5.1.1
distro==1.9.0
docutils==0.20.1
fixtures==4.1.0
frozenlist==1.4.1
future==1.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
keyring==25.2.0
loguru==0.7.2
lxml==5.2.1
magic-filter==1.0.12
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
nh3==0.2.17
nose==1.3.7
openai==1.23.3
pbr==6.0.0
pillow==10.3.0
pkginfo==1.10.0
pycparser==2.22
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.17.2
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
readme_renderer==43.0
redis==5.0.4
requests==2.31.0
requests-mock==1.12.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
sniffio==1.3.1
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
tenacity==8.2.3
testtools==2.7.1
tornado==6.4
tqdm==4.66.2
twine==5.0.0
typing_extensions==4.11.0
urllib3==2.2.1
win32-setctime==1.1.0
yarl==1.9.4
zipp==3.18.1
